+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 * Add stack data structure, implemented in Python.
+* Add queue data structure, implemented in Python.
 
 ## [0.1.0] - 2020-05-03
 ### Added
@@ -15,4 +16,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Add merge sort algorithm, implemented in Python.
 * Add quick sort algorithm, implemented in Python.
 * Add timsort algorithm, implemented in Python.
-* Add sorting alogrithms topic to README.md file.
+* Add sorting algorithms topic to README.md file.