+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.2.0] - 2020-05-03
+### Added
 * Add stack data structure, implemented in Python.
 * Add queue data structure, implemented in Python.
 * Add deque data structure, implemented in Python.